BRIDGE-1317 Only create a notification for scheduled activities when configured to do so.
